sql/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Emanuele Guidotti <emanuele.guidotti@unine.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
bornrule = "^0.2.0"
sqlalchemy = "^2.0.14"
psycopg2-binary = "^2.9.6"
uvicorn = "^0.22.0"
fastapi = "^0.95.2"
pandas = "^1.1.5"
python-dotenv = "0.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
